+++ mod.ts
@@ -1,3 +1,2 @@
 export * from "./fetch_wrapper.ts";
-export * from "./header_utils.ts";
 export type { ExtendedRequestInit } from "./extended_request_init.ts";